chore(protocol): move WorkflowInstanceRelated interface
