+++ engine/src/main/java/io/zeebe/engine/state/ZeebeState.java
@@ -36,8 +36,8 @@ import io.zeebe.engine.state.message.WorkflowInstanceSubscriptionState;
 import io.zeebe.msgpack.UnpackedObject;
 import io.zeebe.protocol.Protocol;
 import io.zeebe.protocol.record.intent.Intent;
-import io.zeebe.protocol.record.intent.WorkflowInstanceRelated;
 import io.zeebe.protocol.record.intent.WorkflowInstanceRelatedIntent;
+import io.zeebe.protocol.record.value.WorkflowInstanceRelated;
 import java.util.function.Consumer;
 import org.slf4j.Logger;
