+++ protocol-impl/src/main/java/io/zeebe/protocol/impl/record/value/job/JobRecord.java
@@ -25,8 +25,8 @@ import io.zeebe.msgpack.property.StringProperty;
 import io.zeebe.msgpack.spec.MsgPackHelper;
 import io.zeebe.protocol.impl.encoding.MsgPackConverter;
 import io.zeebe.protocol.impl.record.UnifiedRecordValue;
-import io.zeebe.protocol.record.intent.WorkflowInstanceRelated;
 import io.zeebe.protocol.record.value.JobRecordValue;
+import io.zeebe.protocol.record.value.WorkflowInstanceRelated;
 import io.zeebe.protocol.record.value.job.Headers;
 import io.zeebe.util.buffer.BufferUtil;
 import java.util.Map;
