+++ protocol-impl/src/main/java/io/zeebe/protocol/impl/record/value/message/WorkflowInstanceSubscriptionRecord.java
@@ -23,7 +23,7 @@ import io.zeebe.msgpack.property.LongProperty;
 import io.zeebe.msgpack.property.StringProperty;
 import io.zeebe.protocol.impl.encoding.MsgPackConverter;
 import io.zeebe.protocol.impl.record.UnifiedRecordValue;
-import io.zeebe.protocol.record.intent.WorkflowInstanceRelated;
+import io.zeebe.protocol.record.value.WorkflowInstanceRelated;
 import io.zeebe.protocol.record.value.WorkflowInstanceSubscriptionRecordValue;
 import io.zeebe.util.buffer.BufferUtil;
 import java.util.Map;
